Initial commit
These are old files, part of an experiment. They need cleaning and
commenting. There are too many references for me to remember after so
long, but I will try to locate and list them, when I get time.

USE_USER_MODE_SCHEDULING
This makes the program use UMS, but it seems to not work perfectly at
the moment.

USE_NEW_THREAD_POOL
This make the program use the New Threading API. This does appear to
work properly.

Both are in the file for comparison purposes.